Source code of the game hangman
This is a basic game that I developed for a project. Terminal will make
a prompt and ask you weather you want to enter a word or should a word
be selected at random by the computer or weather you want to add a word
to the wordlist. It will then give you number of guesses to guess the
word If you did, you won and if you didn’t, you loose.